remove views from base class
