+++ tests/TestBaseClassWeb.php
@@ -33,21 +33,6 @@ class TestBaseClassWeb extends TestBaseClass
     protected $webPort = 4444;
 
     /**
-     * @var array
-     */
-    protected $views;
-
-    /**
-     * @var array
-     */
-    protected $adminViews;
-
-    /**
-     * @var array
-     */
-    protected $adminSurveyViews;
-
-    /**
      * @var WebDriver $webDriver
      */
     protected $webDriver;
@@ -61,8 +46,6 @@ class TestBaseClassWeb extends TestBaseClass
 
         $capabilities = DesiredCapabilities::phantomjs();
         $this->webDriver = RemoteWebDriver::create("http://localhost:{$this->webPort}/", $capabilities);
-        $this->adminViews = require __DIR__."/data/views/adminViews.php";
-        $this->adminSurveyViews = require __DIR__."/data/views/adminSurveyViews.php";
 
     }
