+++ tests/data/views/adminBaseViews.php
